% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-MetabolomicsSet-IO.R
\name{as_ExpressionSet}
\alias{as_ExpressionSet}
\title{Convert to an mSet object to ExpressionSet}
\usage{
as_ExpressionSet(object)
}
\description{
Convert any mSet class object to the Biobase's ExpressionSet
class.
}
\seealso{
\code{\link[Biobase]{ExpressionSet}}
}
