% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-MicrobiomeSet.R
\name{as_phyloseq}
\alias{as_phyloseq}
\title{convert a MicrobiomeSet object to phyloseq}
\usage{
as_phyloseq(mset)
}
\arguments{
\item{mset}{a \code{\link{MicrobiomeSet-class}} object}
}
\description{
convert a MicrobiomeSet object to phyloseq
}
\author{
Chenghao Zhu
}
