% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-MetabolomicsSet-IO.R
\name{assign_lipid_class}
\alias{assign_lipid_class}
\title{Assign lipid class based on lipid annotation name}
\usage{
assign_lipid_class(x)
}
\arguments{
\item{x}{Character vector of the annotated lipid feature names. Must be in
a LipidBlast style}
}
\value{
Character vector fo lipid class.
}
\description{
This function assign lipid class to each lipid feature
according to their annotation names. The annotations need to be in a
LipidBlast style. The lipid class returned is also in a LipidBlast style.
}
\author{
Chenghao Zhu
}
