% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-MetabolomicsSet-IO.R
\name{calibrate_lipidomics_wcmc}
\alias{calibrate_lipidomics_wcmc}
\title{Calculate concentration by calibrating from internal standards}
\usage{
calibrate_lipidomics_wcmc(object, class, cid, ESI)
}
\arguments{
\item{object}{a \code{\link{LipiomicsSet-class}} object}
}
\value{
a \code{\link{LipidomicsSet-class}} object
}
\description{
Generic function for the \code{\link{LipidomcisSet-class}}
object to calculate the concentration of each feature. The object must have
a valid internal-standard slot in the experiment_data. It must also have
the sample_volumn in ul in the experiment_data (slot name: sample_volumn_ul).
}
\details{
The concentration of each feature is calculated as:
\deqn{Conc_{i,j} = \frac{Intensity_{i,j}}{Intensity_{i,k}} \times \frac{Spiked_{k}}{Sample Vol_{i}}}
Where the \eqn{Conc_{i,j}} is the calculated concentration of feature j in
samplei. The \eqn{Intensity_{i,j}} is the raw MS intensity of feature j in
samplei. The \eqn{Intensity_{i,k}} is the raw MS intensity of standard k,
which has the same class as feature j, in sample i. The \eqn{Spiked_{k}} is
the spiked amount in sample k. And the \eqn{Sample Vol_{i}} is the sample
volumn used in the beginning of processing.

The returned concentration is in the unit of ug/ml
}
\author{
Chenghao Zhu
}
