% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-IO.R
\name{collapse_QC}
\alias{collapse_QC}
\title{Collapse QC samples from a MetabolomcisSet object}
\usage{
collapse_QC(object, qc_names)
}
\arguments{
\item{object}{A \code{\link{MetabolomicsSet-class}} object.}

\item{qc_names}{A character vector indicates the names of the QC samples.
Must be from the sampleNames of the object.}
}
\value{
A \code{\link{MetabolomcisSet-class}} object
}
\description{
Calculate the mean, standard deviation, and coefficient of
variance of the QC and put into the feature_data slot. The QC samples are
then removed from conc_table and sample_table.
}
\author{
Chenghao Zhu
}
