% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-mSet.R
\docType{class}
\name{conc_table-class}
\alias{conc_table-class}
\title{S4 class for concentration table}
\description{
This S4 class stores the concentration table from a quantitative experiment.
Inherit from the matrix class. Samples should be in columns and features in
rows. The data should also have sample IDs as column names and feature IDs
as row names.
}
\seealso{
\code{\link{conc_table}}
}
\author{
Chenghao Zhu
}
