% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-assessor.R
\docType{methods}
\name{conc_table<-}
\alias{conc_table<-}
\title{Assign a conc_table slot to a mSet object}
\usage{
\S4method{conc_table}{mSet}(object) <- value
}
\arguments{
\item{object}{A \code{\link{mSet-class}} object}

\item{value}{A \code{\link{conc_table-class}} object}
}
\value{
a \code{\link{mSet-class}} object
}
\description{
This function assign a new \code{\link{conc_table-class}} to
the conc_table slot of a \code{\link{mSet-class}} object.
}
