% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-mSet.R
\docType{methods}
\name{dim,mSet-method}
\alias{dim,mSet-method}
\title{Get dimemsions of the conc_table of a mSet object}
\usage{
\S4method{dim}{mSet}(x)
}
\arguments{
\item{x}{A \code{\link{mSet-class}} or derived class object.}
}
\value{
A integer vector with length of 2, with the first indicates the
number of features, and the second number of samples
}
\description{
Get the dimensions of the conc_table of a mSet object
}
\seealso{
\code{\link{mSet-class}}, \code{\link{conc_table-class}},
\code{\link{MetabolomicsSet}}
}
