% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-assessor.R
\docType{methods}
\name{experiment_data<-}
\alias{experiment_data<-}
\title{Assign a experiment_data slot to a mSet object}
\usage{
\S4method{experiment_data}{mSet}(object) <- value
}
\arguments{
\item{object}{A \code{\link{mSet-class}} object}

\item{value}{A \code{\link{experiment_data-class}} object}
}
\value{
a \code{\link{mSet-class}} object
}
\description{
This function assign a new \code{\link{experiment_data-class}} to
the experiment_data slot of a \code{\link{mSet-class}} object.
}
