% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-assessor.R
\docType{methods}
\name{experiment_data}
\alias{experiment_data}
\title{Access the experiment_data slot from a mSet object}
\usage{
\S4method{experiment_data}{mSet}(object)
}
\arguments{
\item{object}{A \code{\link{mSet-class}} object}
}
\value{
A \code{\link{ExperimentData-class}} object
}
\description{
This is a accessor function for the experiment_data slot from a
\code{\link{mSet-class}} object.
}
