% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-mSet.R
\docType{class}
\name{ExperimentData-class}
\alias{ExperimentData-class}
\title{Deprecated A virtual S4 class for experiment data}
\description{
This class is deprecated in the future
Quantitative experiments such as metabolomics and proteomics often have
additional information beside experiment result. This S4 class is to store
those information such as analysis institute, instrument, resuspension
volumn, and so on.

This is a virtual class and can not be called directly, but only can be
inherited from.
}
\seealso{
\code{\link{MetabolomicsExperimentData-class}}
}
\author{
Chenghao Zhu
}
