% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-MetabolomicsSet-IO.R
\name{export_excle}
\alias{export_excle}
\title{Export mSet object into a Excle sheet}
\usage{
export_excle(object, file, single_sheet = TRUE,
  conc_sheet_name = "intensity matrix",
  samp_sheet_name = "sample info", feat_sheet_name = "features info")
}
\arguments{
\item{object}{A \code{\link{mSet-class}} or derived object.}

\item{file}{A character string indicates the output path.}

\item{single_sheet}{A boolean value, whether write all data slots into a
single sheet or seprate sheets.}

\item{conc_sheet_name, samp_sheet_name, feat_sheet_name}{Character values, with
each indicating the sheet name of each data slot.}
}
\value{
A Excle spreadsheet
}
\description{
This function allows users to export a mSet object to an Excle
spreadsheet. The output is a spreadsheet with feature data on the left,
sample table on the top and transposed, and the conc table in the middle.

This function uses the xlsx pacakge to write the data into Excle format.
Please make sure the package is pre-installed before using it.
}
\author{
Chenghao Zhu
}
