% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-mSet.R
\docType{class}
\name{feature_data-class}
\alias{feature_data-class}
\title{S4 class for feature data.}
\description{
S4 class for storing the feature information from a quantitative experiment.
Inherits from the data.frame class. Rows should be features and columns
should be all the variables. Feature IDs should be row names, and should
match the row IDs of the \code{\link{conc_table-class}}.
}
\seealso{
\code{\link{feature_data}}
}
\author{
Chenghao Zhu
}
