% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-assessor.R
\docType{methods}
\name{feature_data<-}
\alias{feature_data<-}
\title{Assign a feature_data slot to a mSet object}
\usage{
\S4method{feature_data}{mSet}(object) <- value
}
\arguments{
\item{object}{A \code{\link{mSet-class}} object}

\item{value}{A \code{\link{feature_data-class}} object}
}
\value{
a \code{\link{mSet-class}} object
}
\description{
This function assign a new \code{\link{feature_data-class}} to
the feature_data slot of a \code{\link{mSet-class}} object.
}
