% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-mSet.R, R/Methods-assessor.R
\docType{methods}
\name{feature_data}
\alias{feature_data}
\alias{feature_data,mSet-method}
\title{Construct or access a feature_data object}
\usage{
feature_data(object)

\S4method{feature_data}{mSet}(object)
}
\arguments{
\item{object}{This is a data frame of the feature data from a quantitative
experiment to as a constructor. Features should be in rows. And a mSet
object as an assessor.}
}
\value{
A \code{\link{feature_table-class}} object
}
\description{
This is the method to either access the feature data slot from a
\code{\link{mSet-class}} object, or construct a
\code{\link{feature-data-class}} object from a data frame. When use it as an
assessing method, a \code{\link{feature_data-class}} object
is returned. When use it as a construction method, a matrix is required.
}
\seealso{
\code{\link{feature_data-class}}
}
\author{
Chenghao Zhu
}
