% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-MetabolomicsSet-IO.R
\name{filter_by_cv}
\alias{filter_by_cv}
\title{Filter features by QC CV}
\usage{
filter_by_cv(object, cv, cid)
}
\arguments{
\item{object}{A \code{\link{MetabolomicsSet-class}} object}

\item{cv}{A character string, indicates the column name of feature data,
that has the coefficient of variance for each feature.}

\item{cid}{A character string, indicates the column name of feature data,
that has the chemical identifier for each feature. The chemical identifier
must be unique to each feature. Some standard chemical identifier are PubMed
ID, Lipidmaps ID, etc.}
}
\value{
A \code{\link{MetabolomicsSet-class}} object
}
\description{
This function deals with the situation where features are
detetcted in both positive and negative ESI mode. The qc_cv is used as a
reference and the one with a lower cv is kept, and the other one is dropped.
}
\author{
Chenghao Zhu
}
