% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-GlycomicsSet.R
\docType{class}
\name{GlycomicsSet-class}
\alias{GlycomicsSet-class}
\title{S4 class to store glycomics experiment data}
\description{
This is a S4 class inherits from the virtual
\code{\link{mSet-class}} to store the glycomics data. This class acts
as a data container for the glycomics data. Once data is cleaned and
passed into the class, it is ready to do statistical analysis and
visualization.
}
\section{Slots}{

\describe{
\item{\code{conc_table}}{A \code{\link{conc_table-class}} object that stores the
concentration information from the experiment. The column names should be
the feature IDs, and the rownames should be the row IDs. This should be a
numeric matrix.}

\item{\code{sample_table}}{A \code{\link{sample_table-class}} object that stores the
sample meta-data information. The row names should be sample IDs and
should match the column names of the conc_table.}

\item{\code{feature_data}}{A \code{\link{feature_data-class}} object that stores the
feature infromation during the experiment. The row names should be feature
IDs and should match the row names of the conc_table.}

\item{\code{experiment_data}}{A list contains additional experiment information.}
}}

\author{
Chenghao Zhu
}
