% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-GlycomicsSet.R
\name{GlycomicsSet}
\alias{GlycomicsSet}
\title{Construct a GlycomicsSet object}
\usage{
GlycomicsSet(conc_table = NULL, sample_table = NULL,
  feature_data = NULL, experiment_data = NULL)
}
\arguments{
\item{conc_table}{A \code{\link{conc_table-class}} object that stores the
concentration information from the experiment. The column names should be
the feature IDs, and the rownames should be the row IDs. This should be a
numeric matrix.}

\item{sample_table}{A \code{\link{sample_table-class}} object that stores the
sample meta-data information. The row names should be sample IDs and
should match the column names of the conc_table.}

\item{feature_data}{A \code{\link{feature_data-class}} object that stores the
feature infromation during the experiment. The row names should be feature
IDs and should match the row names of the conc_table.}

\item{experiment_data}{A list contains additional experiment information.}
}
\description{
This is the method to construct a
\code{\link{GlycomicsSet-class}} object.
}
\author{
Chenghao Zhu
}
