% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-MetabolomicsSet-IO.R
\name{import_wcmc_excel}
\alias{import_wcmc_excel}
\title{Import West Coast Metabolomics Center data spreadsheet to a
MetabolomicsSet object}
\usage{
import_wcmc_excel(file, sheet, conc_range, sample_range, feature_range,
  InChIKey = NULL, experiment_type = "metabolomics",
  institute = NULL)
}
\arguments{
\item{file}{A character string to the path of the file.}

\item{sheet}{A character string indicates the name of the sheet.}

\item{conc_range}{A character string indicatss the cell range of
\code{\link{conc_table}}, like "B3:D87".}

\item{sample_range}{A character string indicates the cell range of the
\code{\link{sample_table-class}}.}

\item{feature_range}{A character string indicates the cell range of the
\code{\link{feature_data-class}}.}
}
\value{
A \code{\link{MetabolomcisSet-class}} object
}
\description{
This funciton reads the West Coast Metabolomics Center style
data spreadsheet into a MetabolomicsSet object, by specifying the range of
the conc_table, sample_table, and feature_data.
}
\author{
Chenghao Zhu
}
