% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LipidomicsSet.R
\name{lipid_name_formater}
\alias{lipid_name_formater}
\title{Lipid annotation formater for lipidblast}
\usage{
lipid_name_formater(x)
}
\arguments{
\item{x}{character vector of lipid names.}
}
\value{
a character vector
}
\description{
A formater for lipid names for wcmc lipidblast annotations.
Lipid names are forced to be in a style of "PC 36:2 o", as lipid class
abbreviations followed by fatty acid length and number of double bonds. An
additional "d" in the end means a SM or Ceramide, and a "p" or "o" means a
plasmonyl-lipid.
}
