% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-MetabolomicsSet.R
\docType{class}
\name{LipidomicsExperimentData-class}
\alias{LipidomicsExperimentData-class}
\title{S4 class to store experiment data for metabolomics experiment.}
\description{
This is a S4 class inherits from the
\code{\link{MetabolomicsExperimentData-class}}. It can be thought as a
special subclass of the MetabolomicsExperimentData, only for lipidomics
experiment. It's required for the \code{\link{LipidomicsSet-class}} object.
}
\section{Slots}{

\describe{
\item{\code{institute}}{character indicates the institute where the experiment was
conducted}

\item{\code{experiment_type}}{character indicate what type of experiment it is, for
example, lipidomics.}

\item{\code{instument}}{character indicates the instument that the samples were ran}

\item{\code{resuspension_volumn_ul}}{numeric value, the volumn that samples were
resuspended, in ul.}

\item{\code{injection_volumn_ul}}{numeric value, the volumn injected into the
instrument, in ul.}

\item{\code{data_precessing_software}}{character indicates the software the data
processing was done.}

\item{\code{internal_standards}}{a data frame objec. This slot is very strict. It
must have at list 4 columns: InChIKey, class, spike_amt, and spike_unit,
stores the InChIKey, class of the lipid, spiked amount, and the unit,
correspondingly. It also must have all the internal standards appear in the
conc_table slot.}

\item{\code{conc_table_unit}}{character, the conc_table value unit}
}}

\seealso{
\code{\link{LipidomicsSet-class}},
\code{\link{MetabolomicsSet-class}},
\code{\link{MetabolomicsExperimentData-class}},
\code{\link{calibrate}}
}
\author{
Chenghao Zhu
}
