% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-MetabolomicsSet.R
\docType{class}
\name{MetabolomicsExperimentData-class}
\alias{MetabolomicsExperimentData-class}
\title{Deprecated S4 class to store experiment data for metabolomics experiment.}
\description{
This class is deprecated in the future.
This is a S4 class inherits from the virtual class
\code{\link{ExperimentData-class}}. The purpose of this class is to store
all the additional experiment information from a metabolomics experiment.
This class has a list like structure. It contains five slots and each is
either a character or numeric value.
}
\section{Slots}{

\describe{
\item{\code{institute}}{character indicates the institute where the experiment was
conducted}

\item{\code{instument}}{character indicates the instument that the samples were ran}

\item{\code{resuspension_volumn_ul}}{numeric value, the volumn that samples were
resuspended, in ul.}

\item{\code{injection_volumn_ul}}{numeric value, the volumn injected into the
instrument, in ul.}

\item{\code{data_precessing_software}}{character indicates the software the data
processing was done.}
}}

\author{
Chenghao Zhu
}
