% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-MetabolomicsSet.R
\name{MetabolomicsExperimentData}
\alias{MetabolomicsExperimentData}
\title{Construction method for MetabolomicsExperimentData-class}
\usage{
MetabolomicsExperimentData(institute = NULL, instrument = NULL,
  resuspension_volumn_ul = NULL, injection_volumn_ul = NULL,
  data_processing_software = NULL)
}
\arguments{
\item{institute}{character indicates the institute where the experiment was
conducted}

\item{resuspension_volumn_ul}{numeric value, the volumn that samples were
resuspended, in ul.}

\item{injection_volumn_ul}{numeric value, the volumn injected into the
instrument, in ul.}

\item{instument}{character indicates the instument that the samples were ran}

\item{data_precessing_software}{character indicates the software the data
processing was done.}
}
\description{
The construction method to build a
\code{\link{MetabolomicsExperimentData-class}} object.
}
\author{
Chenghao Zhu
}
