% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-MicrobiomeSet.R
\docType{class}
\name{MicrobiomeExperimentData-class}
\alias{MicrobiomeExperimentData-class}
\title{Deprecated S4 class to store the experiment data of a micrbiome experiment}
\description{
This class defines the experiement_slot of the
\code{\link{MicrobiomeSet-class}}. This class inherits from the
\code{\link{ExperiemntData-class}}
}
\section{Slots}{

\describe{
\item{\code{experiment_type}}{character value indicates the type of experiemnt
(16S, 18S, etc.)}

\item{\code{processing_software}}{character value indicates the software used for
data processing/otu clustering (DADA2, qiime2, etc)}
}}

\seealso{
\code{\link{MicorbiomeExperimentData}}
\code{\link{MicrobiomeSet-class}}
\code{\link{ExperimentData-class}}
}
\author{
Chenghao Zhu
}
