% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-MicrobiomeSet.R
\name{MicrobiomeExperimentData}
\alias{MicrobiomeExperimentData}
\title{Construct a MicrobiomeExperimentData object}
\usage{
MicrobiomeExperimentData(experiemnt_type = NULL,
  processing_software = NULL)
}
\arguments{
\item{processing_software}{character value indicates the software used for
data processing/otu clustering (DADA2, qiime2, etc)}

\item{experiment_type}{character value indicates the type of experiemnt
(16S, 18S, etc.)}
}
\description{
The main constructor for the
\code{\link{MicrobiomeExperimentData-class}}
}
\seealso{
\code{\link{MicorbiomeExperimentData-class}}
\code{\link{MicrobiomeSet-class}}
\code{\link{ExperimentData-class}}
}
\author{
Chenghao Zhu
}
