% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-MicrobiomeSet.R
\name{MicrobiomeSet}
\alias{MicrobiomeSet}
\title{Main constructor for MicrobiomeSet class}
\usage{
MicrobiomeSet(conc_table = NULL, sample_table = NULL,
  feature_data = NULL, experiment_data = NULL)
}
\arguments{
\item{conc_table}{A \code{\link{conc_table-class}} object that stores the
concentration information from the experiment. The column names should be
the feature IDs, and the rownames should be the row IDs. This should be a
numeric matrix. This slot is equivalent to the OTU table.}

\item{sample_table}{A \code{\link{sample_table-class}} object that stores the
sample meta-data information. The row names should be sample IDs and
should match the column names of the conc_table.}

\item{feature_data}{A \code{\link{feature_data-class}} object that stores the
feature infromation during the experiment. The row names should be feature
IDs and should match the row names of the conc_table. The feature_data must
be all character and only stores the taxonomy inforamtion
(kingdom to species).}

\item{experiment_data}{A \code{\link{MetabolomicsExperimentData-class}}
object contains additional experiment information.}
}
\description{
This is the main construction method to create a
\code{\link{MicrobiomeSet}} object.
}
\seealso{
\code{\link{MicrobiomeSet-class}} \code{\link{mSet-class}}
\code{\link{MetabolomicsExperimentData-class}}
}
\author{
Chenghao Zhu
}
