% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-mSet.R
\docType{class}
\name{mSet-class}
\alias{mSet-class}
\title{A virtual S4 class to store an quantitative experiment data.}
\description{
This is a virtual S4 class to store the entire dataset from a quantitative
experiment, such as metabolomics and proteomics experiments.

This is a virtual class so it can only be inherited from, but not be
constructed directly. You can either use the classes the inherits this class
defined by this package, or you can define your own class and inherits it to
use some of its features.

The classes that inherits from the virtual mSet class are: the
\code{\link{MetabolomicsSet-class}} and the \code{\link{ProteomicsSet-class}}
.

The mSet class and all the classes that inherits from it should contain at
least the four slots that were discussed below.
}
\section{Slots}{

\describe{
\item{\code{conc_table}}{A \code{\link{conc_table-class}} object that stores the
concentration information from the experiment. The column names should be
the feature IDs, and the rownames should be the row IDs. This should be a
numeric matrix.}

\item{\code{sample_table}}{A \code{\link{sample_table-class}} object that stores the
sample meta-data information. The row names should be sample IDs and
should match the column names of the conc_table.}

\item{\code{feature_data}}{A \code{\link{feature_data-class}} object that stores the
feature infromation during the experiment. The row names should be feature
IDs and should match the row names of the conc_table.}

\item{\code{experiment_data}}{A list contains additional untabular experimental data}
}}

\author{
Chenghao Zhu
}
