% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-stats.R
\name{mSet_deseq}
\alias{mSet_deseq}
\title{Run negative bionomial bistribution based differential expression
analysis with the DESeq2 package}
\usage{
mSet_deseq(object, design, result)
}
\arguments{
\item{object}{\code{\link{mSet-class}} object. The conc_table must be count
data.}

\item{design}{A design matrix. Can be created using the
\code{\link{model.matrix}} funciton.}

\item{result}{integer or string indicates the coefficient to build the
results table.}
}
\description{
Run differential expression analysis base on negative binomial
distribution using the DESeq2 pacakge. The DESeq2 pacakge must be installed.
See \code{\link{DESeq}} for more detail
}
\seealso{
\code{\link{DESeq}}
}
\author{
Chenghao Zhu
}
