% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-stats.R
\name{mSet_limma}
\alias{mSet_limma}
\title{Run linear model with the limma package}
\usage{
mSet_limma(object, design, transform = function(x) {     log2(x + 1) },
  coef, p.value)
}
\arguments{
\item{object}{A \code{\link{mSet-class}} object.}

\item{design}{A design matrix. Can be created using the
\code{\link{model.matrix}} funciton.}

\item{transform}{A transform function to apply. The defaul is log2
transformation.}

\item{coef}{The number of coefficient to return by topTable.
See \code{\link{topTable}}.}
}
\value{
A data.frame. See \code{\link{topTable}}
}
\description{
This is a function that fits linear models for each single
features in a \code{\link{mSet-class}} object. It calls the function
\code{\link{lmFit}}, followed by \code{\link{eBayes}} for t tests.
The result is returned by \code{\link{topTable}}. A model matrix is
requred to be passed to the lmFit function.
}
\author{
Chenghao Zhu
}
