% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-multiSet.R
\docType{class}
\name{MultiExperimentData}
\alias{MultiExperimentData}
\alias{MultiExperimentData}
\title{S4 class to store experiment data for MultiSet}
\usage{
MultiExperimentData(experiment_type)
}
\arguments{
\item{experiment_type}{A character string indicates the type of experiement.}
}
\description{
This is the S4 class for stroing experiment data for the
\code{\link{MultiSet-class}} object.

The constructor for \code{\link{MultiExperimentData-class}}.
See the MultiExperimentData-class for more detail.
}
\seealso{
\code{\link{MultiSet-class}}
}
\author{
Chenghao Zhu

Chenghao Zhu
}
