% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-multiSet.R
\name{MultiSet-class}
\alias{MultiSet-class}
\alias{MultiSet}
\title{S4 class for mutiple experiment types}
\usage{
MultiSet(conc_table = NULL, sample_table = NULL, feature_data = NULL,
  experiment_data = NULL)
}
\description{
The MultiSet class is a flexible data container for different
type of experiment data, such as clinical values, dietary data, or any
biochemical assies. Now it simply inherits from the virtual
\code{\link{mSet-class}} without any modification or restriction. For more
detail, please see the mSet help document.
}
\seealso{
\code{\link{mSet-class}}, \code{\link{MultiSet}}
}
\author{
Chenghao Zhu
}
