% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-multiSet.R
\docType{class}
\name{MultxSet-class}
\alias{MultxSet-class}
\title{S4 class for mutiple experiment types}
\description{
The MultxSet class is a flexible data container for different
type of experiment data, such as clinical values, dietary data, or any
biochemical assies. Now it simply inherits from the virtual
\code{\link{mSet-class}} without any modification or restriction. For more
detail, please see the mSet help document.
}
\seealso{
\code{\link{mSet-class}}, \code{\link{MultxSet}}
}
\author{
Chenghao Zhu
}
