% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LipidomicsSet.R
\name{mz2molwt}
\alias{mz2molwt}
\title{calculate molecular weight from m/z}
\usage{
mz2molwt(species, mz)
}
\arguments{
\item{species}{character value of the adduct ion species}

\item{mz}{numeric value of the m/z}
}
\description{
This function calculate the molecular weight according to the
m/z and adduct ion species.
}
