% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LipidomicsSet.R
\name{nFattyAcyls}
\alias{nFattyAcyls}
\title{Get the number of fatty acyls of a given lipid annotation.}
\usage{
nFattyAcyls(x)
}
\arguments{
\item{x}{character. The lipid annotation name. Must be formatted by the
\code{\link{lipid_name_formater}} function.}
}
\description{
This function returns the number of fatty acyls of a given
lipid annotation name. For example, the number of fatty acyl of lipid PC 34:2
is 2. The lipid annotation names input must be the ouput of
\code{\link{lipid_name_formater}} function.
}
\author{
Chenghao Zhu
}
