% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-mSet.R
\docType{methods}
\name{nfeatures,mSet-method}
\alias{nfeatures,mSet-method}
\title{Get number of features from a mSet object}
\usage{
\S4method{nfeatures}{mSet}(x)
}
\arguments{
\item{x}{A \code{\link{mSet-class}} or derived class object}
}
\value{
A integer value
}
\description{
Get the number of features from a given \code{\link{mSet-class}}
or derived classes.
}
\seealso{
\code{\link{mSet-class}}, \code{\link{conc_table-class}},
\code{\link{feature_data-class}}, \code{\link{MetabolomicsSet-class}}
}
