% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-mSet.R
\docType{methods}
\name{nsamples,mSet-method}
\alias{nsamples,mSet-method}
\title{Get number of samples from a mSet object}
\usage{
\S4method{nsamples}{mSet}(x)
}
\arguments{
\item{x}{A \code{\link{mSet-class}} or derived class object}
}
\value{
A integer value
}
\description{
Get the number of samples from a given \code{\link{mSet-class}}
or derived classes.
}
\seealso{
\code{\link{mSet-class}}, \code{\link{conc_table-class}},
\code{\link{sample_table-class}}, \code{\link{MetabolomicsSet-class}}
}
