% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-plot.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Create a heatmap of an mSet object}
\usage{
plot_heatmap(object, colSideBar = NULL, rowSideBar = NULL,
  scale = "sample", scale.fun = "scale", ...)
}
\arguments{
\item{object}{an mSet object}

\item{colSideBar}{string, the sample}

\item{scale}{string, must be one of sample, feature, or none.}

\item{scale.fun, }{string, must be one either scale or absolute_scale}
}
\description{
Create a heatmap using an mSet object. This function requires the
zheatmap package.
}
\examples{

mset = subset_features(lipid, 1:25)
plot_heatmap(lipid)
plot_heatmap(lipid, colSideBar = "Treatment")
plot_heatmap(lipid, rowsSideBar = "class")

}
\seealso{
\code{\link[zheatmap]{zheatmap}}
}
