% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-plot.R
\name{plot_median_hist}
\alias{plot_median_hist}
\title{Histogram of medians of all features after scaling.}
\usage{
plot_median_hist(object, bins = 30, fill = "steelblue",
  density = TRUE, density_color = "black")
}
\arguments{
\item{object}{A \code{\link{MetabolomicsSet-class}} object.}

\item{bins}{Integer}

\item{fill}{Character, the color fo fill the columns.}

\item{density}{Boolean, whether to plot a density curve}

\item{density_color}{Character, the color of density curve.}
}
\description{
This function makes a histogram of the medians of all features
after scaling into z-scores. It is useful to check the overall normality
of a given dataset.
}
\examples{
data(lipid)
plot_median_hist(lipid)
}
