% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-plot.R
\name{plot_normality_boxplot}
\alias{plot_normality_boxplot}
\title{Plot normality boxplot of each individual feature}
\usage{
plot_normality_boxplot(object, sort = T)
}
\arguments{
\item{object}{A \code{\link{MetabolomicsSet-class}} object.}

\item{sort}{A boolean variable indicating whether arrange the features according
to their medians.}
}
\value{
A ggplot object
}
\description{
This funciton first scale each feature using the \code{\link{scale}}
function, then make a boxplot of the z-score of each feature. This is usful
to check the overall normality of the dataset
}
\examples{
data(lipid)
lipid = subset_features(lipid, 1:10)
plot_normality_bxoplot(lipid)
}
