% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-ProteomicsSet.R
\name{ProteomicsExperimentData}
\alias{ProteomicsExperimentData}
\title{Construction method for the S4 class ProteomicsExperimentData}
\usage{
ProteomicsExperimentData(institute = NULL)
}
\arguments{
\item{institute}{Character indicating the institute or lab where the experiment was conducted.}
}
\description{
The construction method to build a
\code{\link{ProteomicsExperimentData-class}} object.
}
