% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-mSet.R
\docType{class}
\name{sample_table-class}
\alias{sample_table-class}
\title{S4 class for sample metadata.}
\description{
S4 class for storing sample metadata information. Inherits from the
data.frame class. Rows should be samples and columns should be all the
variables. Sample IDs should be row names, and should match the column IDs
of the \code{\link{conc_table-class}}.
}
\seealso{
\code{\link{sample_table}}
}
\author{
Chenghao Zhu
}
