% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-assessor.R
\docType{methods}
\name{sample_table<-}
\alias{sample_table<-}
\title{Assign a sample_table slot to a mSet object}
\usage{
\S4method{sample_table}{mSet}(object) <- value
}
\arguments{
\item{object}{A \code{\link{mSet-class}} object}

\item{value}{A \code{\link{sample_table-class}} object}
}
\value{
a \code{\link{mSet-class}} object
}
\description{
This function assign a new \code{\link{sample_table-class}} to
the sample_table slot of a \code{\link{mSet-class}} object.
}
