% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-mSet.R, R/Methods-assessor.R
\docType{methods}
\name{sample_table}
\alias{sample_table}
\alias{sample_table,mSet-method}
\title{Construction or access a sample_table object}
\usage{
sample_table(object)

\S4method{sample_table}{mSet}(object)
}
\arguments{
\item{object}{This is a data frame of the sample metadata from a quantitative
 experiment as a constructor. Samples should be in rows. And a mSet object
as an assessor.}
}
\value{
A \code{\link{smaple_table-class}} object
}
\description{
This is the method to either access the sample table slot from a
\code{\link{mSet-class}} object, or construct a
\code{\link{sample_table-class}} object from a data frame. When use it as an
assessing method, a \code{\link{sample_table-class}} object
is returned. When use it as a construction method, a matrix is required.
}
\seealso{
\code{\link{sample_table-class}}
}
\author{
Chenghao Zhu
}
