% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-mSet.R
\docType{methods}
\name{subset_samples,mSet,character-method}
\alias{subset_samples,mSet,character-method}
\alias{subset_samples}
\alias{subset_samples,mSet,logical-method}
\alias{subset_samples}
\alias{subset_samples,mSet,integer-method}
\alias{subset_samples}
\alias{subset_features,mSet,character-method}
\alias{subset_features}
\alias{subset_features,mSet,logical-method}
\alias{subset_features}
\alias{subset_features,mSet,integer-method}
\alias{subset_features}
\title{Subset an mSet object}
\usage{
\S4method{subset_samples}{mSet,character}(x, samples)

\S4method{subset_samples}{mSet,logical}(x, samples)

\S4method{subset_samples}{mSet,integer}(x, samples)

\S4method{subset_features}{mSet,character}(x, features)

\S4method{subset_features}{mSet,logical}(x, features)

\S4method{subset_features}{mSet,integer}(x, features)
}
\arguments{
\item{x}{An \code{\link{mSet-class}} or derived object}

\item{samples}{The samples to subset. Can be character, logical, or integers.}

\item{features}{The features to subset. Can be character, logical, or
integers.}
}
\value{
An \code{\link{mSset-class}} or derived object
}
\description{
The subset_samples() and subset_features() functions gets a
subset of an mSet object either by samples or features. The two functions
can either take character vectors of sample/feature names, logical values,
or integer indices to specify the samples or features to subset.
}
