% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LipidomicsSet.R
\name{summarize_ACL}
\alias{summarize_ACL}
\title{Calculate the average carbon chain length}
\usage{
summarize_ACL(object, name, class)
}
\arguments{
\item{object}{a \code{\link{LipidomicsSet-class}} object}

\item{name}{character. The name of the feature variable that contains the
annotation name. The annotation name must be formatted by the
\code{\link{lipid_name_formater}} function}

\item{class}{character. The name of the feature variable that contains the
lipid class.}
}
\description{
Calculate the average carbon chain length of each lipid class.
The input must be a \code{\link{LipidomicsSet-class}} object.
}
\author{
Chenghao Zhu
}
