% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-mSet.R
\name{summarize_features}
\alias{summarize_features}
\title{Summarize mSet by a feature variable}
\usage{
summarize_features(object, feature_var)
}
\arguments{
\item{object}{An \code{\link{mSet-class}} object}

\item{feature_var}{An character variable, must from the column names of the
feature_data slot. Length must be 1.}
}
\value{
An \code{\link{mSet-class}} object
}
\description{
This function takes the conc_table of a mSet object, groups
features by a given feature variable, calculate the sum of each feature in
each group, and return a summarized mSet object.
}
\author{
Chenghao Zhu
}
