% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-LipidomicsSet.R
\name{summarize_lipid_ratios}
\alias{summarize_lipid_ratios}
\title{Calculate lipid classes ratios}
\usage{
summarize_lipid_ratios(object, name, class)
}
\arguments{
\item{object}{a \code{\link{LipidomicsSet-class}} object}

\item{name}{character. The name of the feature variable that contains the
annotation name. The annotation name must be formatted by the
\code{\link{lipid_name_formater}} function}

\item{class}{character. The name of the feature variable that contains the
lipid class.}
}
\description{
Calculate ratios between different lipid classes. The input
must be a \code{\link{LipidomicsSet-class}} object. The lipid class ratios
calculated includes PC/LPC, PE/LPE, CE/free cholesterol, TG/DG, SM / Cer,
CE/TG, surface lipid / core lipid, PC / surface lipid, and SM / surface lipid.
}
\author{
Chenghao Zhu
}
