% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-mSet.R
\name{summarize_samples}
\alias{summarize_samples}
\title{Summarize mSet by sample variables}
\usage{
summarize_samples(object, sample_var)
}
\arguments{
\item{object}{An \code{\link{mSet-class}} object}

\item{sample_var}{An character variable, must from the column names of the
sample_data slot.}
}
\value{
An \code{\link{mSet-class}} object
}
\description{
This function takes the conc_table of a mSet object, groups
samples by given feature variables, calculate the mean of each fsample in
each group, and return a summarized mSet object.
}
\author{
Chenghao Zhu
}
