% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-mSet.R
\name{transform_by_sample}
\alias{transform_by_sample}
\title{Transform an mSet object by sample}
\usage{
transform_by_sample(object, fun, ...)
}
\arguments{
\item{object}{An \code{\link{mSet-class}} or derived class object.}

\item{fun}{A function to apply.}

\item{...}{Arguments to pass to the function/}
}
\value{
An \code{\link{mSet-class}} or derived class object
}
\description{
Transform the conc_table slot of an \code{\link{mSet-class}}
object, one sample at a time. This is similar to MARGIN = 2 in the
\code{\link{apply}} function.
}
\seealso{
\code{\link{mSet-class}}, \code{\link{transform_by_features}}
}
