% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-multiSet.R
\name{MultiSet}
\alias{MultiSet}
\alias{MultiSet}
\title{S4 class for mutiple experiment types}
\usage{
MultiSet(conc_table = NULL, sample_table = NULL, feature_data = NULL,
  experiment_data = NULL)

MultiSet(conc_table = NULL, sample_table = NULL, feature_data = NULL,
  experiment_data = NULL)
}
\arguments{
\item{conc_table}{a \code{\link{conc_table-class}} object}

\item{sample_table}{a \code{\link{sample_table-class}} object}

\item{feature_data}{a \code{\link{feature_data-class}} object}

\item{experiment_data}{a \code{\link{experiment_data-class}} object. The
experiment_data slot in MultiSet is not defined so it can not be used yet.}
}
\description{
The MultiSet class is a flexible data container for different
type of experiment data, such as clinical values, dietary data, or any
biochemical assies. Now it simply inherits from the virtual
\code{\link{mSet-class}} without any modification or restriction. For more
detail, please see the mSet help document.

The main constructor to creat a \code{\link{MultiSet-class}}
object.
}
\seealso{
\code{\link{mSet-class}}, \code{\link{MultiSet}}

\code{\link{MultiSet-class}}, \code{\link{mSet-class}}
}
\author{
Chenghao Zhu

Chenghao Zhu
}
