% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-plot.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Make a boxplot of a feature}
\usage{
plot_boxplot(object, x, feature, rows, cols, line, color, ...)
}
\arguments{
\item{object}{An \code{\link{mSet-class}} object}

\item{x}{A character string indicating the x variable. Must be one of the
sample data variables.}

\item{feature}{A character indicating the feature name to plot. Must be one
of the feature names.}

\item{rows}{A character string indicating the variable defines faceting
groups on columns. Must be one of the sample data variables.}

\item{cols}{A character string indicating the variable defines faceting
groups on columns. Must be one of the sample data variables.}

\item{line}{A character string indicating the variable used to draw lines
between points. Must be one of the sample data variables.}

\item{color}{A character string indicating the variable defines the color
of lines or points. Must be one of the sample data variables.}

\item{...}{Other arguments see \code{\link{ggboxplot}}}
}
\description{
This function uses the ggplot2 package to generate a boxplot from a givin
\code{\link{mSet-class}} object by specifying a feature id. The feature id
must come from the feature names
}
\author{
Chenghao Zhu
}
