% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-mSet.R
\docType{methods}
\name{sampleNames,mSet-method}
\alias{sampleNames,mSet-method}
\alias{sampleNames<-,mSet-method}
\alias{sampleNames<-}
\title{Get or set the sample names of a mSet object}
\usage{
\S4method{sampleNames}{mSet}(x)

\S4method{sampleNames}{mSet}(x) <- value
}
\arguments{
\item{x}{A \code{\link{mSet-class}} or derived class object}

\item{value}{A character vector. The length must equal to the number of
samples of the object.}
}
\value{
A character vector.
}
\description{
Get or set the sample names from a given \code{\link{mSet-class}}
or derived classes.
}
\seealso{
\code{\link{mSet-class}}, \code{\link{conc_table-class}},
\code{\link{sample_table-class}}, \code{\link{MetabolomicsSet-class}}
}
